# Perf: use term hashmap in fastfield